Handle non 'article's when parsing  pubmed xml
As per #63 -- including non 'article' records in a pubmed
xml used to lead to unexepected behaviour in parse_pubmed_xml.

As or this commit we now return an empty record for any non-article
record, and warn the user that this is what we are doing.

This change necessitated  a slight tweak to the pubmed_article
object and the associated print function.