For http/1.1 for https requests
At present, some NCBI requests using http v2 fail. This commit adds code
to force the use of http/1.1 for all requests, and to warn users that
may be setting the http version option for some reason that his is
occuring.

Thanks to @hammer for identifying the problem and proposing the
solution. This closes #127